Add a "testutils" module and move common test code there
