+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'rhino-root'
-include 'rhino', 'rhino-engine', 'rhino-tools', 'rhino-xml', 'rhino-all', 'examples', 'tests', 'benchmarks'
+include 'rhino', 'rhino-engine', 'rhino-tools', 'rhino-xml', 'rhino-all', 'examples', 'testutils', 'tests', 'benchmarks'
