+++ tests/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     implementation project(':rhino-tools')
     implementation project(':rhino-xml')
 
+    testImplementation project(':testutils')
     testImplementation("com.tngtech.archunit:archunit-junit4:1.3.0") {
         exclude group: 'junit'
     }
